requirements.txt
Flask~=1.1.2
requests~=2.24.0
rfc3339-validator~=0.1.2
pytest-cov~=2.10.0
ecdsa~=0.15
jsonschema~=3.2.0
pytz~=2020.1
parsedatetime~=2.6
python-dateutil~=2.8.1
pylint~=2.5.3
requests-mock~=1.8.0
pytest-mock~=3.3.1